+++ README.md
@@ -4,8 +4,6 @@
 ![GitHub](https://img.shields.io/github/license/ing-bank/ginerr)
 ![GitHub go.mod Go version](https://img.shields.io/github/go-mod/go-version/ing-bank/ginerr)
 
-**[❗ 🚨 Click here for version 2 🚨 ❗](./v2)**
-
 Sending any error back to the user can pose a [big security risk](https://owasp.org/www-community/Improper_Error_Handling).
 For this reason we developed an error registry that allows you to register specific error handlers
 for your application. This way you can control what information is sent back to the user.
@@ -15,9 +13,20 @@ You can register errors in 3 ways:
 - By value of string errors
 - By defining the error name yourself
 
+## 👷 V2 migration guide
+
+V2 of this library changes the interface of all the methods to allow contexts to be passed to handlers. This
+allows you to add additional data to the final response.
+
+The interface changes are as follows.
+
+- `RegisterErrorHandler` and all its variants take a context as a first parameter in the handler, allowing you to pass more data to the response
+- `RegisterErrorHandler` and all its variants require the callback function to return `(int, any)` instead of `(int, R)`, removing the unnecessary generic
+- Both `NewErrorResponse` and `NewErrorResponseFrom` take a context as a first parameter, this could be the request context but that's up to you
+
 ## ⬇️ Installation
 
-`go get github.com/ing-bank/ginerr`
+`go get github.com/ing-bank/ginerr/v2`
 
 ## 📋 Usage
 
@@ -26,7 +35,7 @@ package main
 
 import (
 	"github.com/gin-gonic/gin"
-	"github.com/ing-bank/ginerr"
+	"github.com/ing-bank/ginerr/v2"
 	"net/http"
 )
 
@@ -43,7 +52,7 @@ type Response struct {
 }
 
 func main() {
-	handler := func(myError *MyError) (int, Response) {
+	handler := func(ctx context.Context, myError *MyError) (int, any) {
 		return http.StatusInternalServerError, Response{
 			Errors: map[string]any{
 				"error": myError.Error(),
@@ -58,7 +67,7 @@ func main() {
 
 func handleGet(c *gin.Context) {
 	err := &MyError{}
-	c.JSON(ginerr.NewErrorResponse(err))
+	c.JSON(ginerr.NewErrorResponse(c.Request.Context(), err))
 }
 ```
